change int to long int

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -1,4 +1,3 @@
-#include <stdio.h>
 #include "main.h"
 
 /**
@@ -6,7 +5,7 @@
  * @n: number as argument
  */
 
-void print_number(int n)
+void print_number(long int n)
 {
 	int num = n, tenth_power = 1;
